+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 - [Bugfix: fix reactor concurrent bug when get instances & fix spring-retry fuse not working bug](https://github.com/Tencent/spring-cloud-tencent/pull/774)
 - [Standardize the name of method to create bean.](https://github.com/Tencent/spring-cloud-tencent/pull/779)
 - [set reporter.enabled default to true](https://github.com/Tencent/spring-cloud-tencent/pull/784)
+- [fix:fix nacos enabled condition bugs.](https://github.com/Tencent/spring-cloud-tencent/pull/793)